+++ hedera-mirror-test/src/test/java/com/hedera/mirror/test/e2e/acceptance/steps/EstimateFeature.java
@@ -19,7 +19,6 @@ package com.hedera.mirror.test.e2e.acceptance.steps;
 import static com.hedera.mirror.test.e2e.acceptance.client.TokenClient.TokenNameEnum.FUNGIBLE;
 import static com.hedera.mirror.test.e2e.acceptance.steps.AbstractFeature.ContractResource.ESTIMATE_GAS;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.ADDRESS_BALANCE;
-import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.BLAKE2F;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.CALL_CODE_TO_CONTRACT;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.CALL_CODE_TO_EXTERNAL_CONTRACT_FUNCTION;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.CALL_CODE_TO_EXTERNAL_CONTRACT_FUNCTION_VIEW;
@@ -32,13 +31,8 @@ import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.Contra
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.DEPLOY_CONTRACT_VIA_CREATE_OPCODE;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.DEPLOY_CONTRACT_VIA_CREATE_TWO_OPCODE;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.DESTROY;
-import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.ECADD;
-import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.ECMUL;
-import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.ECPAIRING;
-import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.ECRecover_CALL;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.GET_GAS_LEFT;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.GET_MOCK_ADDRESS;
-import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.IDENTITY;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.IERC20_TOKEN_APPROVE;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.IERC20_TOKEN_ASSOCIATE;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.IERC20_TOKEN_DISSOCIATE;
@@ -47,13 +41,10 @@ import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.Contra
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.MESSAGE_SENDER;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.MESSAGE_SIGNER;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.MESSAGE_VALUE;
-import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.MODEXP;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.MULTIPLY_NUMBERS;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.NESTED_CALLS_LIMITED;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.NESTED_CALLS_POSITIVE;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.REENTRANCY_CALL_ATTACK;
-import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.RIPEMD;
-import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.SHA2_CALL;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.STATE_UPDATE_OF_CONTRACT;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.STATIC_CALL_TO_CONTRACT;
 import static com.hedera.mirror.test.e2e.acceptance.steps.EstimateFeature.ContractMethods.STATIC_CALL_TO_INVALID_CONTRACT;
@@ -372,126 +363,6 @@ public class EstimateFeature extends AbstractEstimateFeature {
                 new BigInteger("1025"),
                 asAddress(contractSolidityAddress));
         validateGasEstimation(thirdData, NESTED_CALLS_LIMITED, contractSolidityAddress);
-    }
-
-    @Then("I call estimateGas with call to ethereum precompile ecRecover")
-    public void estimateEcRecover() {
-        final var precompileAddress = "0x0000000000000000000000000000000000000001";
-        final var hash = "0x456e9aea5e197a1f1af7a3e85a3212fa4049a3ba34c2289b4c860fc0b0c64ef3";
-        final var v = "000000000000000000000000000000000000000000000000000000000000001c";
-        final var r = "9242685bf161793cc25603c231bc2f568eb630ea16aa137d2664ac8038825608";
-        final var s = "4f8ae3bd7535248d0bd448298cc2e2071e56992d0774dc340c368ae950852ada";
-
-        final var data = hash.concat(v).concat(r).concat(s);
-
-        validateGasEstimation(data, ECRecover_CALL, precompileAddress);
-    }
-
-    @Then("I call estimateGas with call to ethereum precompile SHA2")
-    public void estimateSHA2() {
-        final var precompileAddress = "0x0000000000000000000000000000000000000002";
-        final var data = "0xFF";
-
-        validateGasEstimation(data, SHA2_CALL, precompileAddress);
-    }
-
-    @Then("I call estimateGas with call to ethereum precompile RIPEMD")
-    public void estimateRIPEMD() {
-        final var precompileAddress = "0x0000000000000000000000000000000000000003";
-        final var data = "0xFF";
-
-        validateGasEstimation(data, RIPEMD, precompileAddress);
-    }
-
-    @Then("I call estimateGas with call to ethereum precompile identity")
-    public void estimateIdentity() {
-        final var precompileAddress = "0x0000000000000000000000000000000000000004";
-        final var data = "0xFF";
-
-        validateGasEstimation(data, IDENTITY, precompileAddress);
-    }
-
-    @Then("I call estimateGas with call to ethereum precompile modexp")
-    public void estimateModexp() {
-        final var precompileAddress = "0x0000000000000000000000000000000000000005";
-        final var bsize = "0000000000000000000000000000000000000000000000000000000000000001";
-        final var esize = "0000000000000000000000000000000000000000000000000000000000000001";
-        final var msize = "0000000000000000000000000000000000000000000000000000000000000001";
-        final var b = "08090A0000000000000000000000000000000000000000000000000000000000";
-        final var data = bsize.concat(esize).concat(msize).concat(b);
-
-        validateGasEstimation(data, MODEXP, precompileAddress);
-    }
-
-    @Then("I call estimateGas with call to ethereum precompile ecAdd")
-    public void estimateEcAdd() {
-        final var precompileAddress = "0x0000000000000000000000000000000000000006";
-        final var x1 = "0x0000000000000000000000000000000000000000000000000000000000000001";
-        final var y1 = "0000000000000000000000000000000000000000000000000000000000000002";
-        final var x2 = "0000000000000000000000000000000000000000000000000000000000000001";
-        final var y2 = "0000000000000000000000000000000000000000000000000000000000000002";
-        final var data = x1.concat(y1).concat(x2).concat(y2);
-
-        validateGasEstimation(data, ECADD, precompileAddress);
-    }
-
-    @Then("I call estimateGas with call to ethereum precompile ecMul")
-    public void estimateEcMul() {
-        final var precompileAddress = "0x0000000000000000000000000000000000000007";
-        final var x1 = "0x0000000000000000000000000000000000000000000000000000000000000001";
-        final var x2 = "0000000000000000000000000000000000000000000000000000000000000002";
-        final var s = "0000000000000000000000000000000000000000000000000000000000000002";
-
-        final var data = x1.concat(x2).concat(s);
-
-        validateGasEstimation(data, ECMUL, precompileAddress);
-    }
-
-    @Then("I call estimateGas with call to ethereum precompile ecPairing")
-    public void estimateEcPairing() {
-        final var precompileAddress = "0x0000000000000000000000000000000000000008";
-        final var x1 = "0x2cf44499d5d27bb186308b7af7af02ac5bc9eeb6a3d147c186b21fb1b76e18da";
-        final var y1 = "2c0f001f52110ccfe69108924926e45f0b0c868df0e7bde1fe16d3242dc715f6";
-        final var x2 = "1fb19bb476f6b9e44e2a32234da8212f61cd63919354bc06aef31e3cfaff3ebc";
-        final var y2 = "22606845ff186793914e03e21df544c34ffe2f2f3504de8a79d9159eca2d98d9";
-        final var x3 = "2bd368e28381e8eccb5fa81fc26cf3f048eea9abfdd85d7ed3ab3698d63e4f90";
-        final var y3 = "2fe02e47887507adf0ff1743cbac6ba291e66f59be6bd763950bb16041a0a85e";
-        final var x4 = "0000000000000000000000000000000000000000000000000000000000000001";
-        final var y4 = "30644e72e131a029b85045b68181585d97816a916871ca8d3c208c16d87cfd45";
-        final var x5 = "1971ff0471b09fa93caaf13cbf443c1aede09cc4328f5a62aad45f40ec133eb4";
-        final var y5 = "091058a3141822985733cbdddfed0fd8d6c104e9e9eff40bf5abfef9ab163bc7";
-        final var x6 = "2a23af9a5ce2ba2796c1f4e453a370eb0af8c212d9dc9acd8fc02c2e907baea2";
-        final var y6 = "23a8eb0b0996252cb548a4487da97b02422ebc0e834613f954de6c7e0afdc1fc";
-
-        final var data = x1.concat(y1)
-                .concat(x2)
-                .concat(y2)
-                .concat(x3)
-                .concat(y3)
-                .concat(x4)
-                .concat(y4)
-                .concat(x5)
-                .concat(y5)
-                .concat(x6)
-                .concat(y6);
-
-        validateGasEstimation(data, ECPAIRING, precompileAddress);
-    }
-
-    @Then("I call estimateGas with call to ethereum precompile blake2f")
-    public void estimateBlake2f() {
-        final var precompileAddress = "0x0000000000000000000000000000000000000009";
-        final var rounds = "0x0000000c";
-        final var h =
-                "48c9bdf267e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f3af54fa5d182e6ad7f520e511f6c3e2b8c68059b6bbd41fbabd9831f79217e1319cde05b";
-        final var m =
-                "6162630000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000";
-        final var t = "03000000000000000000000000000000";
-        final var f = "01";
-
-        final var data = rounds.concat(h).concat(m).concat(t).concat(f);
-
-        validateGasEstimation(data, BLAKE2F, precompileAddress);
     }
 
     @Then("I call estimateGas with IERC20 token transfer using long zero address as receiver")
